Adds typesetting status to both host and guest typesetters
